Remove special behavior when set to 0px; set default to 1400px
I think this makes the setting behavior more intuitive, because the slider value now always means what it seems to mean.

Previously, setting to 0px allowed the page to go to its full 1404px width, but that's basically the same as just setting 1400px for the max width.

I set the lower bound at 100px so that a user can't accidentally make the pages disappear entirely.

Note: I know pages in landscape orientation would have a width of 1872px, but I'm going with 1400px for now because:
- This plugin doesn't seem to be parsing landscape notes correctly at the moment
- 1400px is already very wide